requirements.txt
Django==1.11.15
django-debug-toolbar==1.10.1
djangorestframework==3.5.0
python-decouple==3.1
pytz==2018.7
sqlparse==0.2.4
